SWCME -- Enforce strict warning-clean builds for SWCME writers and demos
- annotate CheckedTextFile::print for compiler format checking
- add a literal-output path that preserves percent characters verbatim
- initialize 3-D writer fields and propagate checked evaluation failures
- cancel staged transactions when field generation fails before formatting
- replace implicit narrowing and signed-index conversions with explicit types
- add debug and optimized OUT08 strict-build targets with warnings as errors
- add and register the OUT08 writer-semantics regression test
- include OUT08 in the SMOKE validation profile
- document strict compiler flags, acceptance criteria, and CI commands

+++ srcSEP/swcme/demo1d.cpp
@@ -205,7 +205,11 @@ int main(){
 
   const double rmin = 0.2*AU, rmax = 1.5*AU;
   for (std::size_t i=0;i<N;++i){
-    const double s = (i + 0.5) / double(N);
+    // Convert the discrete index explicitly at the numerical boundary.  This
+    // makes the intentional floating-point grid parameterization visible to
+    // reviewers and lets OUT08 reject any accidental implicit narrowing.
+    const double s = (static_cast<double>(i) + 0.5) /
+                     static_cast<double>(N);
     r[i] = rmin + s*(rmax - rmin);
   }
